Fixed text issue
Distance now displays properly.
Fixed that one weird bug with the spawn points by raising them slightly (This is not the ideal solution but it will work for now)
Increased the size of despawn because there were some cases where slimes went further in the z axis and they wouldn't despawn.